pyproject.toml
[tool.poetry]
name = "opencoverage"
version = "0.0.1"
description = ""
authors = ["vangheem"]

[tool.poetry.dependencies]
python = "^3.8"
pydantic = "^1.5.1"
orjson = "3.3.1"
fastapi = "^0.61.1"
uvicorn = "^0.12.2"
prometheus-client = "^0.8.0"
python-dotenv = "^0.14.0"
asyncom = "^0.3.1"
aiohttp = "^3.7.3"
lxml = "^4.6.2"
psycopg2-binary = "^2.8.6"
unidiff = "^0.6.0"
cryptography = "^3.3.1"
pyjwt = "^2.0.0"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
pytest-asyncio = "^0.12.0"
mypy = "~0.790"
flake8 = "^3.8.1"
isort = "^5.6.4"
black = "^20.8b1"
pytest-rerunfailures = "^9.0"
pytest-cov = "^2.8.1"
async-asgi-testclient = "^1.4.4"
sqlalchemy-stubs = "^0.4"
codecov = "^2.1.11"

[tool.black]
line-length = 90
target-version = ['py38']
include = '\.pyi?$'
exclude = '''

(
  /(
      \.eggs         # exclude a few common directories in the
    | \.git          # root of the project
    | \.hg
    | \.mypy_cache
    | \.tox
    | \.venv
    | _build
    | buck-out
    | build
    | dist
  )/
  | foo.py           # also separately exclude a file named foo.py in
                     # the root of the project
)
'''

[tool.coverage.report]
show_missing = true
skip_covered = true

[tool.coverage.run]
branch = true
source = ["opencoverage"]


[tool.poetry.scripts]
opencoverage = 'opencoverage.commands:run_command'

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
